✨ feat(components & store): add logo & adjust store structure (#5)
* 🌈 style(*): structure mutation

adjust store structure & remove relative import

* ✨ feat(components): add logo

BTW modified the header href from absolutely to relatively
