+++ src/store/index.ts
@@ -1,0 +1,2 @@
+export * from './useCourse';
+export * from './useFailedCount';
